refactoring of ssr
