+++ src/components/FilmDescription.js
@@ -9,8 +9,8 @@ export class FilmDescription extends React.Component {
    return store.dispatch(fetchInfo(match.params.filmId, match.params.type));
   }
 
-  componentDidMount(nextProps) {
-    if(Object.keys(this.props.currentMovie).length === 0) {
+  componentDidMount() {
+    if(this.props.currentMovie.id != this.props.match.params.filmId) {
       this.props.loadMovieInfo(this.props.match.params.filmId, this.props.match.params.type);
     }
   }
