+++ src/routes.js
@@ -1,10 +1,19 @@
 import React from 'react';
-import '../styles/normalize.css';
-import  '../styles/common.css';
-import AppRoot from './app-root';
-import ListResultContainer from '../containers/ListResultContainer';
-import FilmDescriptionContainer from '../containers/FilmDescriptionContainer';
-import NotFound from '../components/NotFound';
+import './styles/normalize.css';
+import  './styles/common.css';
+import { renderRoutes } from "react-router-config";
+import ListResultContainer from './containers/ListResultContainer';
+import FilmDescriptionContainer from './containers/FilmDescriptionContainer';
+import NotFound from './components/NotFound';
+
+
+const AppRoot = ({route}) => {
+  return (
+    <div>
+      {renderRoutes(route.routes)}
+    </div>
+  );
+};
 
 const routes = [
   { component: AppRoot,
